week-4/py-project-1-customer-support/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
autogen==0.3.1
certifi==2024.8.30
charset-normalizer==3.4.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
FLAML==2.3.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
numpy==1.26.4
openai==1.52.2
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
sniffio==1.3.1
termcolor==2.5.0
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
